requirements.txt
pandas==2.2.2
gspread==6.1.2
google-auth==2.33.0
python-dotenv==1.0.1
